Move IO functions form main and utils to dedicated io.c file
