+++ src/signOffchainMessage.c
@@ -1,3 +1,4 @@
+#include "io.h"
 #include "getPubkey.h"
 #include "os.h"
 #include "ux.h"
